refactor(babylon): improve SonarQube code quality and update outdated dependencies
