+++ .github/workflows/ci.yml
@@ -1,9 +1,5 @@
 name: Continuous Integration
 
-env:
-  STORAGE_NAME: ${{ secrets.STORAGE_NAME }}
-  ACCOUNT_SECRET: ${{ secrets.ACCOUNT_SECRET }}
-
 on:
   workflow_dispatch:
   push:
@@ -55,13 +51,15 @@ jobs:
         with:
           enable-cache: true
 
-      - uses: azure/login@v3
+      - name: Azure login
+        uses: azure/login@v3
         with:
           client-id: ${{ secrets.AZURE_CLIENT_ID }}
           tenant-id: ${{ secrets.AZURE_TENANT_ID }}
           subscription-id: ${{ secrets.AZURE_SUBSCRIPTION_ID }}
 
-      - uses: azure/aks-set-context@v4
+      - name: Azure AKS set context
+        uses: azure/aks-set-context@v4
         with:
           cluster-name: "${{ vars.CLUSTER_NAME }}"
           resource-group: "${{ vars.CLUSTER_RESOURCE_GROUP }}"
@@ -71,10 +69,12 @@ jobs:
           uv venv
           uv pip install .
         shell: bash
+
       - name: Run integration tests
         working-directory: tests/integration/
         run: ./test_api_endpoints.sh
         shell: bash
+        
       - name: End-to-end tests
         working-directory: tests/e2e/
         run: ./test_e2e.sh
