+++ Babylon/commands/macro/deploy_workspace.py
@@ -1,23 +1,13 @@
-import subprocess
 from logging import getLogger
 from pathlib import Path as PathlibPath
-from string import Template
 
 from click import echo, style
-from cosmotech_api.models.workspace_create_request import WorkspaceCreateRequest
-from cosmotech_api.models.workspace_security import WorkspaceSecurity
-from cosmotech_api.models.workspace_update_request import WorkspaceUpdateRequest
-from kubernetes import client, utils
-from kubernetes import config as kube_config
-from kubernetes.utils import FailToCreateError
-from yaml import safe_load
 
 from Babylon.commands.api.workspace import get_workspace_api_instance
-from Babylon.commands.macro.deploy import (
-    create_coal_configmap,
-    create_workspace_secret,
-    get_postgres_service_host,
-    update_object_security,
+from Babylon.commands.macro.helpers.workspace import (
+    create_workspace,
+    deploy_postgres_schema,
+    update_workspace,
 )
 from Babylon.utils.credentials import get_keycloak_token
 from Babylon.utils.environment import Environment
@@ -25,175 +15,6 @@ from Babylon.utils.response import CommandResponse
 
 logger = getLogger(__name__)
 env = Environment()
-
-
-def _create_workspace(api_instance, api_section: dict, payload: dict, state: dict):
-    """Create a new workspace and persist its ID in state. Returns False on failure."""
-    logger.info("  [dim]→ No existing workspace ID found. Creating...[/dim]")
-    workspace = api_instance.create_workspace(
-        organization_id=api_section["organization_id"],
-        workspace_create_request=WorkspaceCreateRequest.from_dict(payload),
-    )
-    if workspace is None:
-        logger.error("  [bold red]✘[/bold red] Failed to create workspace")
-        return False
-    logger.info(f"  [bold green]✔[/bold green] Workspace [bold magenta]{workspace.id}[/bold magenta] created")
-    state["services"]["api"]["workspace_id"] = workspace.id
-    return True
-
-
-def _sync_workspace_security(api_instance, api_section: dict, payload: dict):
-    """Synchronise security roles if a security block is present in the payload."""
-    if not payload.get("security"):
-        return True
-    try:
-        logger.info("  [dim]→ Syncing security policies...[/dim]")
-        current_security = api_instance.get_workspace_security(
-            organization_id=api_section["organization_id"], workspace_id=api_section["workspace_id"]
-        )
-        update_object_security(
-            "workspace",
-            current_security=current_security,
-            desired_security=WorkspaceSecurity.from_dict(payload.get("security")),
-            api_instance=api_instance,
-            object_id=[api_section["organization_id"], api_section["workspace_id"]],
-        )
-        return True
-    except Exception as e:
-        logger.error(f"  [bold red]✘[/bold red] Security update failed: {e}")
-        return False
-
-
-def _update_workspace(api_instance, api_section: dict, payload: dict):
-    """Update an existing workspace and sync its security policy. Returns False on failure."""
-    logger.info(f"  [dim]→ Existing ID [bold cyan]{api_section['workspace_id']}[/bold cyan] found. Updating...[/dim]")
-    updated = api_instance.update_workspace(
-        organization_id=api_section["organization_id"],
-        workspace_id=api_section["workspace_id"],
-        workspace_update_request=WorkspaceUpdateRequest.from_dict(payload),
-    )
-    if updated is None:
-        logger.error(f"  [bold red]✘[/bold red] Failed to update workspace {api_section['workspace_id']}")
-        return False
-    if not _sync_workspace_security(api_instance, api_section, payload):
-        return False
-    logger.info(f"  [bold green]✔[/bold green] Workspace [bold magenta]{api_section['workspace_id']}[/bold magenta] updated")
-    return True
-
-
-def _handle_init_job_logs(k8s_job_name: str, schema_name: str, state: dict) -> None:
-    """Fetch init-job logs and update state based on their content."""
-    logs_process = subprocess.run(
-        ["kubectl", "logs", f"job/{k8s_job_name}", "-n", env.environ_id],
-        capture_output=True,
-        text=True,
-    )
-    if logs_process.returncode != 0:
-        logger.error(f" [bold red]✘[/bold red] Failed to retrieve logs for job {k8s_job_name}")
-        logger.debug(f" [bold red]✘[/bold red] Logs retrieval output {logs_process.stdout} {logs_process.stderr}")
-        return
-
-    job_logs = logs_process.stdout or logs_process.stderr
-    if "ERROR" in job_logs or "error" in job_logs:
-        logger.error("  [bold red]✘[/bold red] Schema creation failed inside the container")
-        logger.debug(f"  [bold red]✘[/bold red] Job logs : {job_logs}")
-    elif "already exists" in job_logs:
-        logger.info(f"  [yellow]⚠[/yellow] [dim]Schema [magenta]{schema_name}[/magenta] already exists (skipping creation)[/dim]")
-    else:
-        logger.info(f"  [green]✔[/green] Schema creation [magenta]{schema_name}[/magenta] completed successfully")
-        state["services"]["postgres"]["schema_name"] = schema_name
-
-
-def _wait_and_check_init_job(k8s_job_name: str, schema_name: str, state: dict) -> None:
-    """Wait for the init job to complete, then inspect its logs."""
-    logger.info(f"  [dim]→ Waiting for job [cyan]{k8s_job_name}[/cyan] to complete...[/dim]")
-    wait_process = subprocess.run(
-        ["kubectl", "wait", "--for=condition=complete", "job", k8s_job_name, f"--namespace={env.environ_id}", "--timeout=50s"],
-        capture_output=True,
-        text=True,
-    )
-    if wait_process.returncode != 0:
-        logger.error(f"  [bold red]✘[/bold red] Job {k8s_job_name} did not complete successfully see babylon logs for details")
-        logger.debug(f"  [bold red]✘[/bold red] Job wait output {wait_process.stdout} {wait_process.stderr}")
-        return
-    logger.info("  [dim]→ Checking job logs for errors...[/dim]")
-    _handle_init_job_logs(k8s_job_name, schema_name, state)
-
-
-def _run_schema_init_job(script_path: PathlibPath, mapping: dict, workspace_id: str, schema_name: str, state: dict) -> None:
-    """Apply a single K8s init job from *script_path* and wait for its outcome."""
-    k8s_job_name = f"postgresql-init-{workspace_id}"
-    kube_config.load_kube_config()
-    k8s_client = client.ApiClient()
-
-    with open(script_path, "r") as f:
-        raw_content = f.read()
-
-    yaml_dict = safe_load(Template(raw_content).safe_substitute(mapping))
-    try:
-        utils.create_from_dict(k8s_client, yaml_dict, namespace=env.environ_id)
-        _wait_and_check_init_job(k8s_job_name, schema_name, state)
-    except FailToCreateError as e:
-        for inner_exception in e.api_exceptions:
-            if inner_exception.status == 409:
-                logger.warning(f"  [yellow]⚠[/yellow] [dim]Job [cyan]{k8s_job_name}[/cyan] already exists.[/dim]")
-            else:
-                logger.error(f"  [bold red]✘[/bold red] K8s Error ({inner_exception.status}): {inner_exception.reason}")
-                logger.debug(f"  Detail: {inner_exception.body}")
-    except Exception as e:
-        logger.error("  [bold red]✘[/bold red] Unexpected error please check babylon logs file for details")
-        logger.debug(f"  [bold red]✘[/bold red] {e}")
-
-
-def _deploy_postgres_schema(workspace_id: str, schema_config: dict, api_section: dict, deploy_dir: PathlibPath, state: dict) -> None:
-    """Initialise the PostgreSQL schema and create the associated K8s resources."""
-    db_host = get_postgres_service_host(env.environ_id)
-    logger.info(f"  [dim]→ Initializing PostgreSQL schema for workspace [bold cyan]{workspace_id}[/bold cyan]...[/dim]")
-
-    pg_config = env.get_config_from_k8s_secret_by_tenant("postgresql-config", env.environ_id)
-    api_config = env.get_config_from_k8s_secret_by_tenant("postgresql-cosmotechapi", env.environ_id)
-    if not pg_config or not api_config:
-        return
-
-    schema_name = workspace_id.replace("-", "_")
-    mapping = {
-        "namespace": env.environ_id,
-        "db_host": db_host,
-        "db_port": "5432",
-        "cosmotech_api_database": api_config.get("database-name", ""),
-        "cosmotech_api_admin_username": api_config.get("admin-username", ""),
-        "cosmotech_api_admin_password": api_config.get("admin-password", ""),
-        "cosmotech_api_writer_username": api_config.get("writer-username", ""),
-        "cosmotech_api_reader_username": api_config.get("reader-username", ""),
-        "workspace_schema": schema_name,
-        "job_name": workspace_id,
-    }
-
-    deploy_dir = deploy_dir if isinstance(deploy_dir, PathlibPath) else PathlibPath(deploy_dir)
-    for job in schema_config.get("jobs", []):
-        script_path = deploy_dir / job.get("path", "") / job.get("name", "")
-        if script_path.exists():
-            _run_schema_init_job(script_path, mapping, workspace_id, schema_name, state)
-
-    organization_id = api_section["organization_id"]
-    logger.info(f"  [dim]→ Creating workspace secret for [cyan]{workspace_id}[/cyan]...[/dim]")
-    create_workspace_secret(
-        namespace=env.environ_id,
-        organization_id=organization_id,
-        workspace_id=workspace_id,
-        writer_password=api_config.get("writer-password", ""),
-    )
-    logger.info(f"  [dim]→ Creating CoAL ConfigMap for [cyan]{workspace_id}[/cyan]...[/dim]")
-    create_coal_configmap(
-        namespace=env.environ_id,
-        organization_id=organization_id,
-        workspace_id=workspace_id,
-        db_host=db_host,
-        db_port="5432",
-        db_name=api_config.get("database-name", ""),
-        schema_name=schema_name,
-        writer_username=api_config.get("writer-username", ""),
-    )
 
 
 def deploy_workspace(namespace: str, file_content: str, deploy_dir: PathlibPath) -> bool:
@@ -211,10 +32,10 @@ def deploy_workspace(namespace: str, file_content: str, deploy_dir: PathlibPath)
 
     # --- Deployment Logic ---
     if not api_section["workspace_id"]:
-        if not _create_workspace(api_instance, api_section, payload, state):
+        if not create_workspace(api_instance, api_section, payload, state):
             return CommandResponse.fail()
     else:
-        if not _update_workspace(api_instance, api_section, payload):
+        if not update_workspace(api_instance, api_section, payload):
             return CommandResponse.fail()
 
     # --- PostgreSQL Schema ---
@@ -222,7 +43,7 @@ def deploy_workspace(namespace: str, file_content: str, deploy_dir: PathlibPath)
     spec = content.get("spec") or {}
     schema_config = spec.get("sidecars", {}).get("postgres", {}).get("schema") or {}
     if schema_config.get("create", False):
-        _deploy_postgres_schema(workspace_id, schema_config, api_section, deploy_dir, state)
+        deploy_postgres_schema(workspace_id, schema_config, api_section, deploy_dir, state)
 
     # --- State Persistence ---
     env.store_state_in_local(state)
