--- Babylon/templates/working_dir/terraform_cloud/tfc_variables_create.yaml
@@ -1,24 +1,0 @@
-- key: var_key,
-  value: var_value,
-  description: var_description,
-  # category: var_category,
-  # hcl: var_hcl,
-  # sensitive: var_sensitive
-- key: var_key,
-  value: var_value,
-  description: var_description,
-  # category: var_category,
-  # hcl: var_hcl,
-  # sensitive: var_sensitive
-- key: var_key,
-  value: var_value,
-  description: var_description,
-  # category: var_category,
-  # hcl: var_hcl,
-  # sensitive: var_sensitive
-- key: var_key,
-  value: var_value,
-  description: var_description,
-  # category: var_category,
-  # hcl: var_hcl,
-  # sensitive: var_sensitive
