--- Babylon/templates/working_dir/terraform_cloud/tfc_workspace_create.yaml
@@ -1,10 +1,0 @@
-# workspace_name: name of the workspace
-workspace_name: ""
-# working_directory: working directory of the workspace
-working_directory: ""
-# vcs_identifier: identifier of the vcs associated to the workspace
-vcs_identifier: ""
-# vcs_branch: branch in the vcs used by the workspace
-vcs_branch: ""
-# vcs_oauth_token_id: id to the oauth token used to connect to the vcs
-vcs_oauth_token_id: ""
